Move custom hooks and functions out of allRebates component and into utilities file
