+++ app/client/src/utilities.tsx
@@ -1,4 +1,8 @@
-import { useQueryClient } from "@tanstack/react-query";
+import { useEffect } from "react";
+import { useQueryClient, useQueries } from "@tanstack/react-query";
+import { useSearchParams } from "react-router-dom";
+// ---
+import { serverUrl, getData } from "./config";
 
 export type CsbData = {
   submissionPeriodOpen: {
@@ -44,6 +48,108 @@ export type BapSamData =
   | { results: false; entities: [] }
   | { results: true; entities: BapSamEntity[] };
 
+type BapFormSubmission = {
+  UEI_EFTI_Combo_Key__c: string; // UEI + EFTI combo key
+  CSB_Form_ID__c: string; // MongoDB ObjectId string
+  CSB_Modified_Full_String__c: string; // ISO 8601 date string
+  CSB_Review_Item_ID__c: string; // CSB Rebate ID with form/version ID (9 digits)
+  Parent_Rebate_ID__c: string; // CSB Rebate ID (6 digits)
+  Record_Type_Name__c:
+    | "CSB Funding Request"
+    | "CSB Payment Request"
+    | "CSB Closeout Request";
+  Parent_CSB_Rebate__r: {
+    CSB_Funding_Request_Status__c: string;
+    CSB_Payment_Request_Status__c: string;
+    CSB_Closeout_Request_Status__c: string;
+    attributes: { type: string; url: string };
+  };
+  attributes: { type: string; url: string };
+};
+
+type FormioApplicationSubmission = {
+  [field: string]: unknown;
+  _id: string; // MongoDB ObjectId string
+  state: "submitted" | "draft";
+  modified: string; // ISO 8601 date string
+  data: {
+    [field: string]: unknown;
+    // fields injected upon new draft Application submission creation:
+    last_updated_by: string;
+    hidden_current_user_email: string;
+    hidden_current_user_title: string;
+    hidden_current_user_name: string;
+    bap_hidden_entity_combo_key: string;
+    sam_hidden_applicant_email: string;
+    sam_hidden_applicant_title: string;
+    sam_hidden_applicant_name: string;
+    sam_hidden_applicant_efti: string;
+    sam_hidden_applicant_uei: string;
+    sam_hidden_applicant_organization_name: string;
+    sam_hidden_applicant_street_address_1: string;
+    sam_hidden_applicant_street_address_2: string;
+    sam_hidden_applicant_city: string;
+    sam_hidden_applicant_state: string;
+    sam_hidden_applicant_zip_code: string;
+    // fields set by form definition (among others):
+    applicantUEI: string;
+    applicantEfti: string;
+    applicantEfti_display: string;
+    applicantOrganizationName: string;
+    schoolDistrictName: string;
+  };
+};
+
+type FormioPaymentRequestSubmission = {
+  [field: string]: unknown;
+  _id: string; // MongoDB ObjectId string
+  state: "submitted" | "draft";
+  modified: string; // ISO 8601 date string
+  data: {
+    [field: string]: unknown;
+    // fields injected upon new draft Payment Request submission creation:
+    bap_hidden_entity_combo_key: string;
+    hidden_application_form_modified: string; // ISO 8601 date string
+    hidden_current_user_email: string;
+    hidden_current_user_title: string;
+    hidden_current_user_name: string;
+    hidden_bap_rebate_id: string;
+  };
+};
+
+type FormioCloseOutSubmission = {
+  [field: string]: unknown;
+  _id: string; // MongoDB ObjectId string
+  state: "submitted" | "draft";
+  modified: string; // ISO 8601 date string
+  data: {
+    [field: string]: unknown;
+  };
+};
+
+type BapSubmission = {
+  modified: string | null; // ISO 8601 date string
+  comboKey: string | null; // UEI + EFTI combo key
+  rebateId: string | null; // CSB Rebate ID (6 digits)
+  reviewItemId: string | null; // CSB Rebate ID with form/version ID (9 digits)
+  status: string | null;
+};
+
+export type Rebate = {
+  application: {
+    formio: FormioApplicationSubmission;
+    bap: BapSubmission | null;
+  };
+  paymentRequest: {
+    formio: FormioPaymentRequestSubmission | null;
+    bap: BapSubmission | null;
+  };
+  closeOut: {
+    formio: FormioCloseOutSubmission | null;
+    bap: BapSubmission | null;
+  };
+};
+
 /** Custom hook that returns cached fetched CSB data */
 export function useCsbData() {
   const queryClient = useQueryClient();
@@ -54,6 +160,310 @@ export function useCsbData() {
 export function useBapSamData() {
   const queryClient = useQueryClient();
   return queryClient.getQueryData<BapSamData>(["bap-sam-data"]);
+}
+
+/** Custom hook to fetch submissions from the BAP and Formio */
+export function useSubmissionsQueries() {
+  return useQueries({
+    queries: [
+      {
+        queryKey: ["bap-form-submissions"],
+        queryFn: () => {
+          const url = `${serverUrl}/api/bap-form-submissions`;
+          return getData<BapFormSubmission[]>(url).then((res) => {
+            const submissions = res.reduce(
+              (object, submission) => {
+                const formType =
+                  submission.Record_Type_Name__c === "CSB Funding Request"
+                    ? "applications"
+                    : submission.Record_Type_Name__c === "CSB Payment Request"
+                    ? "paymentRequests"
+                    : submission.Record_Type_Name__c === "CSB Closeout Request"
+                    ? "closeOuts"
+                    : null;
+
+                if (formType) object[formType].push(submission);
+
+                return object;
+              },
+              {
+                applications: [] as BapFormSubmission[],
+                paymentRequests: [] as BapFormSubmission[],
+                closeOuts: [] as BapFormSubmission[],
+              }
+            );
+
+            return Promise.resolve(submissions);
+          });
+        },
+        refetchOnWindowFocus: false,
+      },
+      {
+        queryKey: ["formio-application-submissions"],
+        queryFn: () => {
+          const url = `${serverUrl}/api/formio-application-submissions`;
+          return getData<FormioApplicationSubmission[]>(url);
+        },
+        refetchOnWindowFocus: false,
+      },
+      {
+        queryKey: ["formio-payment-request-submissions"],
+        queryFn: () => {
+          const url = `${serverUrl}/api/formio-payment-request-submissions`;
+          return getData<FormioPaymentRequestSubmission[]>(url);
+        },
+        refetchOnWindowFocus: false,
+      },
+      {
+        queryKey: ["formio-close-out-submissions"],
+        queryFn: () => {
+          const url = `${serverUrl}/api/formio-close-out-submissions`;
+          return getData<FormioCloseOutSubmission[]>(url);
+        },
+        refetchOnWindowFocus: false,
+      },
+    ],
+  });
+}
+
+/**
+ * Custom hook to combine Application form submissions data, Payment Request
+ * form submissions data, and Close-Out form submissions data from both the BAP
+ * and Formio into a single `submissions` object, with the BAP assigned
+ * `rebateId` as the keys.
+ **/
+function useCombinedRebates() {
+  const queryClient = useQueryClient();
+
+  const bapFormSubmissions = queryClient.getQueryData<{
+    applications: BapFormSubmission[];
+    paymentRequests: BapFormSubmission[];
+    closeOuts: BapFormSubmission[];
+  }>(["bap-form-submissions"]);
+
+  const formioApplicationSubmissions = queryClient.getQueryData<
+    FormioApplicationSubmission[]
+  >(["formio-application-submissions"]);
+
+  const formioPaymentRequestSubmissions = queryClient.getQueryData<
+    FormioPaymentRequestSubmission[]
+  >(["formio-payment-request-submissions"]);
+
+  const formioCloseOutSubmissions = queryClient.getQueryData<
+    FormioCloseOutSubmission[]
+  >(["formio-close-out-submissions"]);
+
+  // ensure form submissions data has been fetched from both the BAP and Formio
+  if (
+    !bapFormSubmissions ||
+    !formioApplicationSubmissions ||
+    !formioPaymentRequestSubmissions ||
+    !formioCloseOutSubmissions
+  ) {
+    return {};
+  }
+
+  const rebates: { [rebateId: string]: Rebate } = {};
+
+  /**
+   * Iterate over Formio Application form submissions, matching them with
+   * submissions returned from the BAP, so we can build up each rebate object
+   * with the Application form submission data and initialize Payment Request
+   * form and Close-out Form submission data structure (both to be updated).
+   */
+  for (const formioSubmission of formioApplicationSubmissions) {
+    const bapMatch = bapFormSubmissions.applications.find((bapSub) => {
+      return bapSub.CSB_Form_ID__c === formioSubmission._id;
+    });
+
+    const modified = bapMatch?.CSB_Modified_Full_String__c || null;
+    const comboKey = bapMatch?.UEI_EFTI_Combo_Key__c || null;
+    const rebateId = bapMatch?.Parent_Rebate_ID__c || null;
+    const reviewItemId = bapMatch?.CSB_Review_Item_ID__c || null;
+    const status =
+      bapMatch?.Parent_CSB_Rebate__r?.CSB_Funding_Request_Status__c || null;
+
+    /**
+     * NOTE: If new Application form submissions have been reciently created in
+     * Formio and the BAP's ETL process has not yet run to pickup those new
+     * Formio submissions, all of the fields above will be null, so instead of
+     * assigning the submission's key as `rebateId` (which will be null), we'll
+     * assign it to be an underscore concatenated with the Formio submission's
+     * mongoDB ObjectID – just so each submission object still has a unique ID.
+     */
+    rebates[rebateId || `_${formioSubmission._id}`] = {
+      application: {
+        formio: { ...formioSubmission },
+        bap: { modified, comboKey, rebateId, reviewItemId, status },
+      },
+      paymentRequest: { formio: null, bap: null },
+      closeOut: { formio: null, bap: null },
+    };
+  }
+
+  /**
+   * Iterate over Formio Payment Request form submissions, matching them with
+   * submissions returned from the BAP, so we can set the Payment Request form
+   * submission data.
+   *
+   * NOTE: For there to be any Formio Payment Request form submissions at all,
+   * the BAP's ETL process must be running, as the `hidden_bap_rebate_id` field
+   * of a Payment Request form submission is injected in the creation of a brand
+   * new submission in the `/api/formio-payment-request-submission` POST request
+   * where he BAP Rebate ID (along with other fields) are fetched from the BAP
+   * and then posted to Formio in a new Payment Request form submission.
+   *
+   * That said, if the BAP ETL isn't returning data, we should make sure we
+   * handle that situation gracefully (see NOTE below).
+   */
+  for (const formioSubmission of formioPaymentRequestSubmissions) {
+    const formioBapRebateId = formioSubmission.data.hidden_bap_rebate_id;
+
+    const bapMatch = bapFormSubmissions.paymentRequests.find((bapSub) => {
+      return bapSub.Parent_Rebate_ID__c === formioBapRebateId;
+    });
+
+    const modified = bapMatch?.CSB_Modified_Full_String__c || null;
+    const comboKey = bapMatch?.UEI_EFTI_Combo_Key__c || null;
+    const rebateId = bapMatch?.Parent_Rebate_ID__c || null;
+    const reviewItemId = bapMatch?.CSB_Review_Item_ID__c || null;
+    const status =
+      bapMatch?.Parent_CSB_Rebate__r?.CSB_Payment_Request_Status__c || null;
+
+    /**
+     * NOTE: If the BAP ETL is running, there should be a submission with a
+     * `formioBapRebateId` key for each Formio Payment Request form submission
+     * (as it would have been set in the `formioApplicationSubmissions` loop
+     * above). That said, we should first check that it exists before assigning
+     * the Payment Request data to it, so if the BAP ETL process isn't returning
+     * data, it won't break our app.
+     */
+    if (rebates[formioBapRebateId]) {
+      rebates[formioBapRebateId].paymentRequest = {
+        formio: { ...formioSubmission },
+        bap: { modified, comboKey, rebateId, reviewItemId, status },
+      };
+    }
+  }
+
+  /**
+   * Iterate over Formio Close-Out form submissions, matching them with
+   * submissions returned from the BAP, so we can set the Close-Out form
+   * submission data.
+   */
+  for (const formioSubmission of formioCloseOutSubmissions) {
+    console.log(formioSubmission); // TODO
+  }
+
+  return rebates;
+}
+
+/**
+ * Custom hook that sorts rebates by:
+ * - most recient formio modified date, regardless of form
+ *   (Application, Payment Request, or Close-Out)
+ * - Application submissions needing edits
+ * - selected Applications submissions without a corresponding Payment Request
+ *   submission
+ **/
+function useSortedRebates(rebates: { [rebateId: string]: Rebate }) {
+  return Object.entries(rebates)
+    .map(([rebateId, rebate]) => ({ rebateId, ...rebate }))
+    .sort((r1, r2) => {
+      const mostRecientR1Modified = [
+        Date.parse(r1.application.formio.modified),
+        Date.parse(r1.paymentRequest.formio?.modified || ""),
+        Date.parse(r1.closeOut.formio?.modified || ""),
+      ].reduce((previous, current) => {
+        return current > previous ? current : previous;
+      });
+
+      const mostRecientR2Modified = [
+        Date.parse(r2.application.formio.modified),
+        Date.parse(r2.paymentRequest.formio?.modified || ""),
+        Date.parse(r2.closeOut.formio?.modified || ""),
+      ].reduce((previous, current) => {
+        return current > previous ? current : previous;
+      });
+
+      return mostRecientR2Modified - mostRecientR1Modified;
+    })
+    .sort((r1, _r2) => {
+      const r1ApplicationNeedsEdits = submissionNeedsEdits({
+        formio: r1.application.formio,
+        bap: r1.application.bap,
+      });
+
+      const r1PaymentRequestNeedsEdits = submissionNeedsEdits({
+        formio: r1.paymentRequest.formio,
+        bap: r1.paymentRequest.bap,
+      });
+
+      const r1ApplicationSelected = r1.application.bap?.status === "Accepted";
+
+      const r1ApplicationSelectedButNoPaymentRequest =
+        r1ApplicationSelected && !Boolean(r1.paymentRequest.formio);
+
+      return r1ApplicationNeedsEdits ||
+        r1PaymentRequestNeedsEdits ||
+        r1ApplicationSelectedButNoPaymentRequest
+        ? -1
+        : 0;
+    });
+}
+
+/**
+ * Custom hook that returns sorted rebates, and logs them if 'debug' search
+ * parameter exists.
+ */
+export function useRebates() {
+  const [searchParams] = useSearchParams();
+
+  const combinedRebates = useCombinedRebates();
+  const sortedRebates = useSortedRebates(combinedRebates);
+
+  // log combined 'sortedRebates' array if 'debug' search parameter exists
+  useEffect(() => {
+    if (searchParams.has("debug") && sortedRebates.length > 0) {
+      console.log(sortedRebates);
+    }
+  }, [searchParams, sortedRebates]);
+
+  return sortedRebates;
+}
+
+/**
+ * Determines whether a submission needs edits, based on the BAP status
+ *
+ * NOTE: we can't use the BAP status alone though, because if a submission has
+ * been re-submitted and the BAP hasn't yet run their ETL to pickup the status
+ * change, we need to ensure we properly display the 'submitted' formio status.
+ */
+export function submissionNeedsEdits(options: {
+  formio:
+    | FormioApplicationSubmission
+    | FormioPaymentRequestSubmission
+    | FormioCloseOutSubmission
+    | null;
+  bap: BapSubmission | null;
+}) {
+  const { formio, bap } = options;
+
+  if (!formio) return false;
+
+  /**
+   * The submission has been updated in Formio since the last time the BAP's
+   * submissions ETL process has last succesfully run.
+   */
+  const submissionHasBeenUpdatedSinceLastETL = bap?.modified
+    ? new Date(formio.modified) > new Date(bap.modified)
+    : false;
+
+  return (
+    bap?.status === "Edits Requested" &&
+    (formio.state === "draft" ||
+      (formio.state === "submitted" && !submissionHasBeenUpdatedSinceLastETL))
+  );
 }
 
 /**
